quantum_messenger/quantum_messenger.py
"""
Quantum NFL Internal Messaging System
A secure, quantum-inspired messaging system for internal team communication
"""

import datetime
import jwt
import uuid
from cryptography.fernet import Fernet
from dataclasses import dataclass
from typing import List, Optional, Dict

@dataclass
class QuantumMessage:
    id: str
    sender: str
    recipients: List[str]
    content: str
    quantum_timestamp: float
    entanglement_id: Optional[str] = None
    priority_level: int = 1
    encryption_key: Optional[bytes] = None
    
class QuantumMessenger:
    def __init__(self):
        self.encryption_key = Fernet.generate_key()
        self.cipher_suite = Fernet(self.encryption_key)
        self.message_store: Dict[str, QuantumMessage] = {}
        self.entangled_messages: Dict[str, List[str]] = {}
        
    def create_message(self, sender: str, recipients: List[str], content: str, 
                      priority: int = 1) -> str:
        """Create a new quantum message with entanglement support"""
        message_id = str(uuid.uuid4())
        quantum_time = datetime.datetime.now().timestamp()
        
        # Encrypt the message content
        encrypted_content = self.cipher_suite.encrypt(content.encode())
        
        message = QuantumMessage(
            id=message_id,
            sender=sender,
            recipients=recipients,
            content=encrypted_content.decode(),
            quantum_timestamp=quantum_time,
            priority_level=priority
        )
        
        self.message_store[message_id] = message
        return message_id
        
    def entangle_messages(self, message_ids: List[str]) -> str:
        """Create quantum entanglement between messages"""
        entanglement_id = str(uuid.uuid4())
        self.entangled_messages[entanglement_id] = message_ids
        
        for msg_id in message_ids:
            if msg_id in self.message_store:
                self.message_store[msg_id].entanglement_id = entanglement_id
                
        return entanglement_id
        
    def read_message(self, message_id: str, reader: str) -> Optional[str]:
        """Read a message with quantum state collapse"""
        if message_id not in self.message_store:
            return None
            
        message = self.message_store[message_id]
        if reader not in message.recipients and reader != message.sender:
            return None
            
        # Decrypt the message
        decrypted_content = self.cipher_suite.decrypt(
            message.content.encode()
        ).decode()
        
        return decrypted_content
        
    def get_entangled_messages(self, entanglement_id: str) -> List[QuantumMessage]:
        """Retrieve all messages in an entangled state"""
        if entanglement_id not in self.entangled_messages:
            return []
            
        return [
            self.message_store[msg_id] 
            for msg_id in self.entangled_messages[entanglement_id]
            if msg_id in self.message_store
        ]

class QuantumTeam:
    def __init__(self, team_id: str, members: List[str]):
        self.team_id = team_id
        self.members = set(members)
        self.quantum_state = "active"
        
    def add_member(self, member: str) -> bool:
        if member not in self.members:
            self.members.add(member)
            return True
        return False
        
    def remove_member(self, member: str) -> bool:
        if member in self.members:
            self.members.remove(member)
            return True
        return False

class QuantumMessagingSystem:
    def __init__(self):
        self.messenger = QuantumMessenger()
        self.teams: Dict[str, QuantumTeam] = {}
        self.user_sessions: Dict[str, str] = {}  # user -> session_token
        
    def create_team(self, team_id: str, members: List[str]) -> QuantumTeam:
        """Create a new quantum team"""
        team = QuantumTeam(team_id, members)
        self.teams[team_id] = team
        return team
        
    def send_team_message(self, sender: str, team_id: str, content: str, 
                         priority: int = 1) -> Optional[str]:
        """Send a message to all team members"""
        if team_id not in self.teams:
            return None
            
        team = self.teams[team_id]
        if sender not in team.members:
            return None
            
        return self.messenger.create_message(
            sender=sender,
            recipients=list(team.members),
            content=content,
            priority=priority
        )
        
    def authenticate_user(self, username: str, password: str) -> Optional[str]:
        """Authenticate user and return session token"""
        # In production, implement proper authentication
        token = jwt.encode(
            {"user": username, "exp": datetime.datetime.utcnow() + datetime.timedelta(hours=1)},
            self.messenger.encryption_key,
            algorithm="HS256"
        )
        self.user_sessions[username] = token
        return token

# Example usage
if __name__ == "__main__":
    # Initialize the system
    qms = QuantumMessagingSystem()
    
    # Create teams
    analysis_team = qms.create_team("analysis", ["analyst1", "analyst2", "lead_analyst"])
    dev_team = qms.create_team("development", ["dev1", "dev2", "lead_dev"])
    
    # Authenticate users
    analyst_token = qms.authenticate_user("analyst1", "password")
    
    # Send team message
    message_id = qms.send_team_message(
        sender="analyst1",
        team_id="analysis",
        content="New quantum patterns detected in recent games!",
        priority=2
    )
    
    # Read message
    if message_id:
        content = qms.messenger.read_message(message_id, "analyst2")
        print(f"Message received: {content}")
