quantum_messenger/requirements.txt
cryptography>=41.0.0
pyjwt>=2.8.0
python-dateutil>=2.8.2
pyyaml>=6.0.1
requests>=2.31.0
websockets>=11.0.3
